for #2084, move SetAssignmentsExtractor to update package
