+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/antlr/extractor/impl/dml/update/SetAssignmentsExtractor.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.parse.antlr.extractor.impl.dml;
+package org.apache.shardingsphere.core.parse.antlr.extractor.impl.dml.update;
 
 import com.google.common.base.Optional;
 import org.antlr.v4.runtime.ParserRuleContext;
